lab9 done

+++ ComputerVision/Core/models.h
@@ -11,13 +11,13 @@ class Models
     vector<pair<shared_ptr<Descriptor>, shared_ptr<Descriptor>>> matches;
 public:
     vector<pair<Point, Point>> best;
-    float bestModel[9];
+    double bestModel[9];
 
     Models(vector<shared_ptr<Descriptor>> desc1, vector<shared_ptr<Descriptor>> desc2);
 
-    float *RanSaC(int iterCount, float eps);
+    double *RanSaC(int iterCount, float eps);
 
-    float *Hough(float eps);
+    double *Hough(float eps);
     ~Models();
 };
